init auth module & controller
