+++ modules/node/src/auth/auth.module.ts
@@ -1,0 +1,11 @@
+import { Module } from "@nestjs/common";
+
+import { AuthController } from "./auth.controller";
+import { AuthService } from "./auth.service";
+
+@Module({
+  controllers: [AuthController],
+  exports: [AuthService],
+  providers: [AuthService],
+})
+export class AuthModule {}
